@@ -14,7 +14,9 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+
+publish_to: none
 
 dependencies:
   flutter:
