@@ -3,20 +3,20 @@ description: >-
   A Flutter package that implements Material Design Feature discovery
   to show a description of specific features to new users.
   See https://tinyurl.com/FeatureDiscovery
-version: 0.13.0+1
+version: 0.14.0
 homepage: https://github.com/ayalma/feature_discovery
 
 environment:
   # This should match the Flutter stable SDK constraint.
   # See https://github.com/flutter/flutter/blob/stable/packages/flutter/pubspec.yaml.
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  provider: ^4.0.4
-  shared_preferences: ^0.5.6+1
+  provider: ^5.0.0
+  shared_preferences: ^2.0.3
 
 dev_dependencies:
   flutter_test:
