@@ -34,9 +34,9 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key key, this.title}) : super(key: key);
+  const MyHomePage({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -61,7 +61,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: Text(widget.title!),
         bottom: PreferredSize(
           preferredSize: const Size.fromHeight(80),
           child: Column(
@@ -103,30 +103,32 @@ class _MyHomePageState extends State<MyHomePage> {
               children: <Widget>[
                 const Text(
                     'Also, notice how the pulsing animation is not playing because it is deactivated for this feature.'),
-                FlatButton(
-                    child: Text('Toggle enablePulsingAnimation',
-                        style: Theme.of(context)
-                            .textTheme
-                            .button
-                            .copyWith(color: Colors.white)),
-                    onPressed: () => setState(() {
-                          feature1EnablePulsingAnimation =
-                              !feature1EnablePulsingAnimation;
-                        })),
+                TextButton(
+                  onPressed: () => setState(() {
+                    feature1EnablePulsingAnimation =
+                        !feature1EnablePulsingAnimation;
+                  }),
+                  child: Text('Toggle enablePulsingAnimation',
+                      style: Theme.of(context)
+                          .textTheme
+                          .button!
+                          .copyWith(color: Colors.white)),
+                ),
                 const Text(
                     'Ignore the items below or tap the button to toggle between OverflowMode.clip and OverflowMode.doNothing!'),
-                FlatButton(
-                    child: Text('Toggle overflowMode',
-                        style: Theme.of(context)
-                            .textTheme
-                            .button
-                            .copyWith(color: Colors.white)),
-                    onPressed: () => setState(() {
-                          feature1OverflowMode =
-                              feature1OverflowMode == OverflowMode.clipContent
-                                  ? OverflowMode.ignore
-                                  : OverflowMode.clipContent;
-                        })),
+                TextButton(
+                  onPressed: () => setState(() {
+                    feature1OverflowMode =
+                        feature1OverflowMode == OverflowMode.clipContent
+                            ? OverflowMode.ignore
+                            : OverflowMode.clipContent;
+                  }),
+                  child: Text('Toggle overflowMode',
+                      style: Theme.of(context)
+                          .textTheme
+                          .button!
+                          .copyWith(color: Colors.white)),
+                ),
                 for (int n = 42; n > 0; n--)
                   const Text('Testing clipping (ignore or toggle)',
                       style: TextStyle(backgroundColor: Colors.black)),
@@ -149,24 +151,26 @@ class _MyHomePageState extends State<MyHomePage> {
               children: <Widget>[
                 const Text(
                     'Tap the magnifying glass to quickly scan your compounds'),
-                FlatButton(
-                  padding: const EdgeInsets.all(0),
-                  child: Text('Understood',
-                      style: Theme.of(context)
-                          .textTheme
-                          .button
-                          .copyWith(color: Colors.white)),
+                TextButton(
                   onPressed: () async =>
                       FeatureDiscovery.completeCurrentStep(context),
+                  child: Text(
+                    'Understood',
+                    style: Theme.of(context)
+                        .textTheme
+                        .button!
+                        .copyWith(color: Colors.white),
+                  ),
                 ),
-                FlatButton(
-                  padding: const EdgeInsets.all(0),
-                  child: Text('Dismiss',
-                      style: Theme.of(context)
-                          .textTheme
-                          .button
-                          .copyWith(color: Colors.white)),
+                TextButton(
                   onPressed: () => FeatureDiscovery.dismissAll(context),
+                  child: Text(
+                    'Dismiss',
+                    style: Theme.of(context)
+                        .textTheme
+                        .button!
+                        .copyWith(color: Colors.white),
+                  ),
                 ),
               ],
             ),
@@ -190,15 +194,16 @@ class _MyHomePageState extends State<MyHomePage> {
           description: Column(children: <Widget>[
             const Text(
                 'This is overly long to test OverflowMode.extendBackground. The green circle should be large enough to cover all of the text.'),
-            FlatButton(
-                child: Text('Add another item',
-                    style: Theme.of(context)
-                        .textTheme
-                        .button
-                        .copyWith(color: Colors.white)),
-                onPressed: () => setState(() {
-                      feature3ItemCount++;
-                    })),
+            TextButton(
+              onPressed: () => setState(() {
+                feature3ItemCount++;
+              }),
+              child: Text('Add another item',
+                  style: Theme.of(context)
+                      .textTheme
+                      .button!
+                      .copyWith(color: Colors.white)),
+            ),
             for (int n = feature3ItemCount; n > 0; n--)
               const Text('Testing OverflowMode.extendBackground'),
           ]),
@@ -214,22 +219,22 @@ class _MyHomePageState extends State<MyHomePage> {
 }
 
 class Content extends StatefulWidget {
-  const Content({Key key}) : super(key: key);
+  const Content({Key? key}) : super(key: key);
 
   @override
   _ContentState createState() => _ContentState();
 }
 
 class _ContentState extends State<Content> {
-  GlobalKey<EnsureVisibleState> ensureKey;
-  GlobalKey<EnsureVisibleState> ensureKey2;
+  GlobalKey<EnsureVisibleState>? ensureKey;
+  GlobalKey<EnsureVisibleState>? ensureKey2;
 
   @override
   void initState() {
     ensureKey = GlobalKey<EnsureVisibleState>();
     ensureKey2 = GlobalKey<EnsureVisibleState>();
 
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       FeatureDiscovery.discoverFeatures(
         context,
         const <String>{
@@ -309,8 +314,8 @@ class _ContentState extends State<Content> {
                     return true;
                   },
                   onOpen: () async {
-                    WidgetsBinding.instance.addPostFrameCallback((_) {
-                      ensureKey.currentState.ensureVisible(
+                    WidgetsBinding.instance!.addPostFrameCallback((_) {
+                      ensureKey!.currentState!.ensureVisible(
                         preciseAlignment: 0.5,
                         duration: const Duration(milliseconds: 400),
                       );
@@ -323,8 +328,7 @@ class _ContentState extends State<Content> {
                   contentLocation: ContentLocation.below,
                   child: EnsureVisible(
                     key: ensureKey,
-                    child: RaisedButton(
-                      child: const Text('Start Feature Discovery'),
+                    child: ElevatedButton(
                       onPressed: () {
                         FeatureDiscovery.discoverFeatures(
                           context,
@@ -338,6 +342,7 @@ class _ContentState extends State<Content> {
                           },
                         );
                       },
+                      child: const Text('Start Feature Discovery'),
                     ),
                   ),
                 ),
@@ -359,8 +364,8 @@ class _ContentState extends State<Content> {
                     return true;
                   },
                   onOpen: () async {
-                    WidgetsBinding.instance.addPostFrameCallback((_) {
-                      ensureKey2.currentState.ensureVisible(
+                    WidgetsBinding.instance!.addPostFrameCallback((_) {
+                      ensureKey2!.currentState!.ensureVisible(
                           duration: const Duration(milliseconds: 600));
                     });
                     return true;
@@ -368,16 +373,16 @@ class _ContentState extends State<Content> {
                   description: Column(children: <Widget>[
                     const Text(
                         'You can test OverflowMode.wrapBackground here.'),
-                    FlatButton(
-                        padding: const EdgeInsets.all(0),
-                        child: Text('Add item',
-                            style: Theme.of(context)
-                                .textTheme
-                                .button
-                                .copyWith(color: Colors.white)),
-                        onPressed: () => setState(() {
-                              feature6ItemCount++;
-                            })),
+                    TextButton(
+                      onPressed: () => setState(() {
+                        feature6ItemCount++;
+                      }),
+                      child: Text('Add item',
+                          style: Theme.of(context)
+                              .textTheme
+                              .button!
+                              .copyWith(color: Colors.white)),
+                    ),
                     for (int n = feature6ItemCount; n > 0; n--)
                       const Text('Testing OverflowMode.wrapBackground'),
                   ]),
@@ -416,10 +421,10 @@ class _ContentState extends State<Content> {
               child: FloatingActionButton(
                 backgroundColor: Colors.white,
                 foregroundColor: Colors.blue,
-                child: const Icon(Icons.drive_eta),
                 onPressed: () {
                   print('Floating action button tapped.');
                 },
+                child: const Icon(Icons.drive_eta),
               ),
             ),
           ),
